feat: Refactor to use React Query
