+++ frontend/src/components/ResetModal/ResetModal.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import axios from 'axios';
 import { Modal } from 'antd';
 import { useAppDispatch } from 'hooks';
 import { resetCourses } from 'reducers/coursesSlice';
 import { resetTabs } from 'reducers/courseTabsSlice';
 import { resetPlanner } from 'reducers/plannerSlice';
-import { useSelector } from 'react-redux';
-import { RootState } from 'config/store';
+import { useQuery } from 'react-query';
+import { resetDegree } from 'utils/api/degreeApi';
 
 type Props = {
   open?: boolean;
@@ -17,14 +16,11 @@ type Props = {
 // has no internal "open" state since it becomes difficult to juggle with external buttons
 const ResetModal = ({ open, onOk, onCancel }: Props) => {
   const dispatch = useAppDispatch();
-  const { token } = useSelector((state: RootState) => state.settings);
   const handleOk = async () => {
     dispatch(resetPlanner());
-    try {
-      await axios.post('user/reset', {}, { params: { token } });
-    } catch (err) {
-      // eslint-disable-next-line no-console
-      console.error('Error resetting degree at handleDegreeChange: ' + err);
+    const degreeQuery = useQuery('degree', resetDegree);
+    if (degreeQuery.isError) {
+      console.log('Error while resetting degree');
     }
     dispatch(resetTabs());
     dispatch(resetCourses());
